ban status use case in progress 2

+++ src/users/use-cases/ban-status-change-use-case.ts
@@ -1,12 +1,19 @@
-import { CommentsRepository } from './../comments/comments.repository';
+import { CommentsRepository } from '../../comments/comments.repository';
 import { BlogsRepository } from 'src/blogs/blogs.repository';
-import { UsersRepository } from './../users/users.repository';
+import { UsersRepository } from '../users.repository';
 import { Injectable } from "@nestjs/common";
 import { PostsRepository } from 'src/posts/posts.repository';
 import { UsersDevicesRepository } from 'src/usersDevices/usersDevicesRepository';
+import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
+import { BanUserInputModel } from '../sa.users.controller';
 
-@Injectable()
-export class BanService{
+
+export class BanStatusChangeCommand {
+    constructor(public userId, public banDto: BanUserInputModel){}
+}
+
+@CommandHandler(BanStatusChangeCommand)
+export class BanStatusChangeUseCase implements ICommandHandler<BanStatusChangeCommand>{
     constructor(
         private readonly usersRepository: UsersRepository,
         private readonly usersDevicesRepository: UsersDevicesRepository,
@@ -14,12 +21,31 @@ export class BanService{
         private readonly postsRepository: PostsRepository,
         private readonly commentsRepository: CommentsRepository,){}
 
-        async banUser(userId: string){
-            const user = await this.usersRepository.getUserDBTypeById(userId)
-            user.isBanned = true
-            user.save()
+        async execute(command: BanStatusChangeCommand){
+            const userId = command.userId
+            const isBanned = command.banDto.isBanned
 
-            await this.usersDevicesRepository.deleteAllUserDevicesById(userId)
+            const user = await this.usersRepository.getUserDBTypeById(userId)
+            
+            if(user.isBanned === isBanned){
+                return
+            }
+            if(isBanned === true){
+                user.banReason = command.banDto.banReason
+                await this.usersDevicesRepository.deleteAllUserDevicesById(userId)
+            }
+            user.isBanned = isBanned
+            const userBanResult = await this.usersRepository.saveUser(user)
+
+            const postsBanResult = await this.postsRepository.setBanStatusForPosts(userId, isBanned)
+
+            
+
+
+
+
+
+            
 
             
         }
